move: getItemCardMessages.tsをserviceディレクトリからutilディレクトリに移動
