+++ app/_utils/functions/getItemCardMessages.ts
